[refactor] :recycle: refactor use of utils functions
