+++ src/services/contractSetupService.ts
@@ -4,10 +4,35 @@ import { Logger } from '../utils/logger';
 import { IBlockchain } from '../models/blockchain';
 import { getChatterpayABI } from './bucketService';
 import { getNetworkConfig } from './networkService';
-import { getBundlerUrl, validateBundlerUrl } from '../utils/bundler';
-import { ComputedAddress, computeProxyAddressFromPhone } from './predictWalletService';
+import { validateBundlerUrl } from '../utils/bundler';
+import { setupContractReturnType } from '../types/common';
+import { computeProxyAddressFromPhone } from './predictWalletService';
 
+/**
+ * Returns a valid public Bundler RPC URL from Stackup given a chain id
+ * @param chainId The chain ID Number
+ * @returns {string} (The url)
+ */
+function getBundlerUrl(chainId: number): string {
+  const bundlerUrls: { [key: number]: string | undefined } = {
+    1: 'https://public.stackup.sh/api/v1/node/ethereum-mainnet',
+    11155111: 'https://public.stackup.sh/api/v1/node/ethereum-sepolia',
+    137: 'https://public.stackup.sh/api/v1/node/polygon-mainnet',
+    80001: 'https://public.stackup.sh/api/v1/node/polygon-mumbai',
+    43114: 'https://public.stackup.sh/api/v1/node/avalanche-mainnet',
+    43113: 'https://public.stackup.sh/api/v1/node/avalanche-fuji',
+    10: 'https://public.stackup.sh/api/v1/node/optimism-mainnet',
+    11155420: 'https://public.stackup.sh/api/v1/node/optimism-sepolia',
+    56: 'https://public.stackup.sh/api/v1/node/bsc-mainnet',
+    97: 'https://public.stackup.sh/api/v1/node/bsc-testnet',
+    42161: 'https://public.stackup.sh/api/v1/node/arbitrum-one',
+    421614: process.env.ARBITRUM_SEPOLIA_RPC_URL,
+    8453: 'https://public.stackup.sh/api/v1/node/base-mainnet',
+    84532: 'https://public.stackup.sh/api/v1/node/base-sepolia'
+  };
 
+  return bundlerUrls[chainId] || '';
+}
 
 /**
  * Sets up the necessary contracts and providers for blockchain interaction.
